move HelpFile from gui.help to logic.help
