+++ src/main/java/net/sf/jabref/gui/help/HelpAction.java
@@ -28,6 +28,7 @@ import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.gui.IconTheme;
 import net.sf.jabref.gui.actions.MnemonicAwareAction;
 import net.sf.jabref.gui.desktop.JabRefDesktop;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.l10n.Localization;
 
 import org.apache.commons.logging.Log;
