+++ src/main/java/net/sf/jabref/gui/preftabs/AdvancedTab.java
@@ -28,8 +28,8 @@ import javax.swing.JTextField;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.gui.help.HelpAction;
-import net.sf.jabref.gui.help.HelpFile;
 import net.sf.jabref.gui.remote.JabRefMessageHandler;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.remote.RemotePreferences;
 import net.sf.jabref.logic.remote.RemoteUtil;
