+++ src/main/java/net/sf/jabref/importer/fetcher/CiteSeerXFetcher.java
@@ -25,10 +25,10 @@ import java.util.regex.Pattern;
 
 import javax.swing.JPanel;
 
-import net.sf.jabref.gui.help.HelpFile;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.logic.formatter.bibtexfields.NormalizeNamesFormatter;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.net.URLDownload;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.IdGenerator;
