+++ src/main/java/net/sf/jabref/importer/fetcher/DBLPFetcher.java
@@ -24,10 +24,10 @@ import java.util.List;
 
 import javax.swing.JPanel;
 
-import net.sf.jabref.gui.help.HelpFile;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.importer.fileformat.BibtexParser;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.net.URLDownload;
 import net.sf.jabref.model.DuplicateCheck;
 import net.sf.jabref.model.entry.BibEntry;
