+++ src/main/java/net/sf/jabref/logic/fulltext/ArXiv.java
@@ -16,7 +16,7 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
-import net.sf.jabref.gui.help.HelpFile;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.importer.fetcher.IdBasedFetcher;
 import net.sf.jabref.importer.fetcher.OAI2Fetcher;
 import net.sf.jabref.importer.fetcher.SearchBasedFetcher;
