+++ src/main/java/net/sf/jabref/logic/fulltext/SearchBasedEntryFetcher.java
@@ -22,12 +22,12 @@ import java.util.Objects;
 
 import javax.swing.JPanel;
 
-import net.sf.jabref.gui.help.HelpFile;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.importer.fetcher.EntryFetcher;
 import net.sf.jabref.importer.fetcher.SearchBasedFetcher;
 import net.sf.jabref.logic.fetcher.FetcherException;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.entry.BibEntry;
